Test for root

+++ install.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
-echo " "
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
+
+echo "ADBasher Installer, tested on Kali and Parrot"
 sudo apt update && apt autoremove && apt autoclean 
-# sudo uname -a
+echo "Current OS version is:"
+sudo uname -a
 sudo apt install grc crackmapexec
